Refactoring of the generator module
